factor out viewer into new module
